refactor(ci): split up overlarge run.ts and runProjectsInBulk
